+++ extension-new/utils/config.js
@@ -1,5 +1,5 @@
 var Config = {
-  WS_URL: 'ws://localhost:25429/plugin',
+  WS_URL: 'ws://localhost:37171/plugin',
   RECONNECT_DELAY: 3000,
   DEBUGGER_VERSION: '1.3',
   HEARTBEAT_INTERVAL: 25000,